[AIRFLOW-5918][part of AIRFLOW-5893] Group tests for the Pools command (#6567)
